requirements.txt
asgiref==3.2.10
azure-common==1.1.26
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
beautifulsoup4==4.9.3
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.4
cryptography==3.2.1
dj-database-url==0.5.0
Django==3.1.3
django-appconf==1.0.4
django-bootstrap4==2.2.0
django-cors-headers==3.5.0
django-debug-toolbar==3.1.1
django-filter==2.4.0
django-imagekit==4.0.2
django-storages==1.10.1
djangorestframework==3.11.2
djangorestframework-jwt==1.11.0
easy-thumbnails==2.7.1
gunicorn==20.0.4
httpie==2.2.0
idna==2.10
pilkit==2.0
Pillow==8.0.1
plotly==4.13.0
psycopg2-binary==2.8.6
pycparser==2.20
Pygments==2.7.1
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
retrying==1.3.3
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
urllib3==1.25.10
whitenoise==5.2.0
